refactor: Exception code 인터페이스화 및 도메인 별 분리
