+++ src/main/java/com/spaceclub/form/FormExceptionMessage.java
@@ -1,0 +1,16 @@
+package com.spaceclub.form;
+
+import com.spaceclub.global.exception.ExceptionMessageInterface;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum FormExceptionMessage implements ExceptionMessageInterface {
+
+    FORM_NOT_FOUND("존재하지 않는 폼입니다"),
+    FORM_OPTION_NOT_FOUND("존재하지 않는 폼 옵션 입니다");
+
+    private final String message;
+
+}
