+++ src/main/java/com/spaceclub/form/service/FormService.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.spaceclub.global.ExceptionCode.FORM_NOT_FOUND;
+import static com.spaceclub.form.FormExceptionMessage.FORM_NOT_FOUND;
 
 @Service
 @Transactional(readOnly = true)
