+++ src/main/java/com/spaceclub/global/exception/ControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.spaceclub.global;
+package com.spaceclub.global.exception;
 
 import org.springframework.core.annotation.Order;
 import org.springframework.http.ResponseEntity;
