+++ src/main/java/com/spaceclub/global/exception/ExceptionMessageInterface.java
@@ -1,0 +1,7 @@
+package com.spaceclub.global.exception;
+
+public interface ExceptionMessageInterface {
+
+    String getMessage();
+
+}
